Update README.Remix-Based Simple Storage Smart Contract
This repository contains the code and instructions for Lesson 2 of the Cyfrin Solidity Blockchain Course, focusing on the Ethereum development environment, Remix. The lesson introduces 'SimpleStorage.sol', a Solidity contract exemplifying the basics of smart contract storage on the Ethereum blockchain. The guide walks through how to deploy and interact with this contract using Remix. Additionally, it provides resources for further learning, including a comprehensive blog on working with Remix and video content.